Improvements to File and FileSystem classes
The file is now an enseble of parent file path and file name. the path is the union of these two strings.
Create a new File constructor to fit these features and add methds to the FileSystem class to respect also this "separation" (passing to method a parent path and a file name)